MAINT: interpolate: move pchip to _monotone.py
